pyproject.toml
[tool.poetry]
name = "fastapitesting"
version = "0.0.1"
description = "A repo for practicing testing with FastAPI"
authors = ["Rehket <aalbright425@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
databases = "^0.5.5"
pydantic = "^1.9.0"
httpx = "^0.22.0"
uvicorn = "^0.17.6"
secure = "^0.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
PyJWT = "^2.3.0"
aiosqlite = "^0.17.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-asyncio = "^0.18.3"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# pyproject.toml
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "fastapi_testing/tests",
]
asyncio_mode = "strict"